feat: add env
