+++ .gitignore
@@ -14,3 +14,9 @@ Thumbs.db
 
 # Logs
 npm-debug.log
+# Logs
+npm-debug.log
+
+#dist
+
+dist
