+++ tsconfig.json
@@ -11,7 +11,7 @@
       // "disableReferencedProjectLoad": true,             /* Reduce the number of projects loaded automatically by TypeScript. */
   
       /* Language and Environment */
-      "target": "ES2021",                                  /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */
+      "target": "esnext",
       // "lib": [],                                        /* Specify a set of bundled library declaration files that describe the target runtime environment. */
       // "jsx": "preserve",                                /* Specify what JSX code is generated. */
       // "experimentalDecorators": true,                   /* Enable experimental support for TC39 stage 2 draft decorators. */
@@ -25,7 +25,7 @@
       // "moduleDetection": "auto",                        /* Control what method is used to detect module-format JS files. */
   
       /* Modules */
-      "module": "commonjs",                                /* Specify what module code is generated. */
+      "module": "esnext",
       "rootDir": "./",                                  /* Specify the root folder within your source files. */
       "moduleResolution": "node",                       /* Specify how TypeScript looks up a file from a given module specifier. */
       // "baseUrl": "./",                                  /* Specify the base directory to resolve non-relative module names. */
@@ -49,7 +49,10 @@
       // "emitDeclarationOnly": true,                      /* Only output d.ts files and not JavaScript files. */
       // "sourceMap": true,                                /* Create source map files for emitted JavaScript files. */
       // "outFile": "./",                                  /* Specify a file that bundles all outputs into one JavaScript file. If 'declaration' is true, also designates a file that bundles all .d.ts output. */
-      "outDir": "./dist",                                   /* Specify an output folder for all emitted files. */
+      "outDir": "./dist",  
+      "noEmit": true,
+      "allowImportingTsExtensions": true,
+
       // "removeComments": true,                           /* Disable emitting comments. */
       // "noEmit": true,                                   /* Disable emitting files from a compilation. */
       // "importHelpers": true,                            /* Allow importing helper functions from tslib once per project, instead of including them per-file. */
