+++ webpack.config.js
@@ -1,0 +1,24 @@
+const path = require('path');
+
+module.exports = {
+  entry: './index.ts',
+  output: {
+    path: path.resolve(__dirname, './dist'),
+    filename: 'index.js',
+  },
+  mode: 'production',
+  target: 'node',
+  module: {
+    rules: [
+      {
+        test: /\.(ts|tsx)$/i,
+        loader: 'ts-loader',
+        include: [path.resolve(__dirname, 'src')],
+        exclude: ['/node_modules/'],
+      },
+    ]
+  },
+  resolve: {
+    extensions: ['.tsx', '.ts', '.jsx', '.js', '...'],
+  }, 
+}
